WIP #959 renamed users-store -> users
