+++ src/server/storage/storage.js
@@ -1,7 +1,7 @@
 'use strict';
 var MongoClient = require('mongodb').MongoClient,
     RPCStore = require('../rpc/storage'),
-    UserStore = require('./user-store'),
+    Users = require('./users'),
     Projects = require('./projects'),
     UserActions = require('./user-actions'),
     PublicProjects = require('./public-projects');
@@ -17,7 +17,7 @@ Storage.prototype.connect = function() {
     var mongoURI = process.env.MONGO_URI || process.env.MONGOLAB_URI || 'mongodb://localhost:27017';
     return MongoClient.connect(mongoURI)
         .then(db => {
-            this.users = new UserStore(this._logger, db);
+            this.users = new Users(this._logger, db);
             this.projects = Projects;
             Projects.init(this._logger, db);
             RPCStore.init(this._logger, db);
